requirements.txt
django>=2.2.13, <3.0
djangorestframework>=3.11.0, <4.0
django-filter>=2.3.0, <3.0
django-guardian>=2.3.0, <3.0
pillow>=7.1.2, <8.0
mysqlclient>=1.4.6, >2.0
pymysql>=0.9.3, <1.0
coreapi>=2.3.3, <3.0